Minor update to schemawrap.sh script

+++ scripts/schemawrap.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 
 # Takes all JSON Schema files in the current directory and wraps them into a "definitions" 
-# section of a single schema. Also takes a base schema name and enclosed it in an "allOf"
-# clause at the top level of the schema.
+# section of a single schema. Also takes a base schema and puts it at the top level of the schema.
 
 if [ "$#" -ne 2 ]; then
     echo "Usage: <output_file> <schema_base>"
@@ -12,10 +11,13 @@ fi
 output_file=$1
 schema_base=$2
 
+if [ ! -f ${schema_base_file} ]; then
+    echo "Schema base file not found!"
+    exit 1
+fi
+
 echo "{" > ${output_file}
 echo " \"\$schema\": \"http://json-schema.org/draft-04/schema#\"," >> ${output_file}
-echo " \"description\": \"Schema that describes a template\"," >> ${output_file}
-echo " \"type\": \"object\", " >> ${output_file}
 
 # create "definitions" section"
 echo " \"definitions\": {" >> ${output_file}
@@ -34,7 +36,7 @@ do
 
  while IFS='' read -r line || [[ -n "${line}" ]]; do
   # Replace "file:" with "#/definitions/" in $ref clause and strip ".json" from names
-   echo "    "${line} | sed "s/\"file:/\"#\/definitions\//" | sed "s/\.json\"/\"/" >> ${output_file}
+  echo "    "${line} | sed "s/\"file:/\"#\/definitions\//" | sed "s/\.json\"/\"/" >> ${output_file}
  done < ${file}
 
 done
@@ -42,12 +44,19 @@ done
 # end of "definitions"
 echo "  }," >> ${output_file}
 
+schema_base_file=${schema_base}.json
+
+
+# Strip the opening and closing parenthesis lines from the base schema file
+sed -e '1d' -e '$d' ${schema_base_file} | while IFS='' read -r line || [[ -n "${line}" ]]; do
+ # Replace "file:" with "#/definitions/" in $ref clause and strip ".json" from names
+ echo "    "${line} | sed "s/\"file:/\"#\/definitions\//" | sed "s/\.json\"/\"/" >> ${output_file}
+done 
+ 
 # start of "allOf"
-echo "    \"allOf\": [" >> ${output_file}
-
-echo "      { \"\$ref\": \"#/definitions/${schema_base}\" }" >> ${output_file}
-
-echo "    ]" >> ${output_file}
+#echo "    \"allOf\": [" >> ${output_file}
+#echo "      { \"\$ref\": \"#/definitions/${schema_base}\" }" >> ${output_file}
+#echo "    ]" >> ${output_file}
 #end of "allOf"
 
 echo "}" >> ${output_file}
